Add ‘transaction_version’ to the signed transaction
This allows hardware wallets to know which transactions they can safely
sign.  To reduce transaction size, I reduced it to a ‘u8’ from a ‘u32’.

Fixes #5951.

+++ primitives/version/src/lib.rs
@@ -103,7 +103,7 @@ pub struct RuntimeVersion {
 	/// index.
 	///
 	/// It need *not* change when a new module is added or when a dispatchable is added.
-	pub transaction_version: u32,
+	pub transaction_version: u8,
 }
 
 #[cfg(feature = "std")]
